refactor: move validate operational conditions from compressor train, rename to validate model input
chore: add ModelInputFailureStatus

chore: fix tests
